requirements.txt
pandas>=1.3.0
psycopg2-binary>=2.9.1
vk-api>=11.9.7
Telethon>=1.24.0
python-dotenv>=0.19.1
gspread>=5.0.0
gspread-dataframe>=3.3.0
oauth2client>=4.1.3
markdown-it-py>=2.0.0
pytz>=2021.3
argparse>=1.4.0 